+++ src/jsonSchemas/connectorConfig.ts
@@ -3,7 +3,7 @@
 // npx ts-json-schema-generator -p ./src/jsonSchemas/connectorConfig.ts -o ./src/jsonSchemas/connectorConfig.json -t "ConnectorConfig" --no-top-ref
 
 export interface ConnectorConfig {
-    mode: "debug" | "production";
+    debug: boolean;
 
     database: {
         connectionString: string;